Added Namespace filter for Super Admin
