+++ gx/gx-core-flow/src/main/java/io/graphenee/core/flow/security/GxSecurityPolicyListView.java
@@ -1,5 +1,7 @@
 package io.graphenee.core.flow.security;
 
+import io.graphenee.common.GxAuthenticatedUser;
+import io.graphenee.core.flow.GxUserAccountDashboardUser;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.vaadin.flow.component.HasComponents;
@@ -18,8 +20,6 @@ public class GxSecurityPolicyListView extends GxVerticalLayoutView {
 	@Autowired
 	GxSecurityPolicyList list;
 
-	GxNamespace namespace;
-
 	@Override
 	protected void decorateLayout(HasComponents rootLayout) {
 		rootLayout.add(list);
@@ -27,13 +27,20 @@ public class GxSecurityPolicyListView extends GxVerticalLayoutView {
 
 	@Override
 	public void afterNavigation(AfterNavigationEvent event) {
-		namespace = list.getNamespace();
-		list.initializeWithNamespace(namespace);
+		list.initializeWithNamespace(getNamespace());
 	}
 
 	@Override
 	protected String getCaption() {
 		return "Security Policies";
+	}
+
+	public GxNamespace getNamespace() {
+		GxAuthenticatedUser user = loggedInUser();
+		if(user instanceof GxUserAccountDashboardUser){
+			return ((GxUserAccountDashboardUser)user).getUser().getNamespace();
+		}
+		return null;
 	}
 
 }
